I assigned the variable computerNumber to the number that the computer is going to generate
I assigned the variable userNumber to the number that the user will guess. I set it to 0 because the user has not guessed anything yet.

I created an object for the scanner class and an object for the random class

We set the computerNumber to generate a random number. I placed the number as 101 so that the range was from 1 - 100